requirements.txt
ridi-cms-sdk
flask>=0.10.0